Model 파일로 관리
