+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Person.swift
@@ -1,0 +1,93 @@
+//
+//  Person.swift
+//  ParcelInvoiceMaker
+//
+//  Created by Eunsung on 4/16/24.
+//
+
+import Foundation
+
+struct Address {
+    private let address: String
+    init(value: String) throws {
+        guard !value.isEmpty else {
+            throw errorAddress.empty
+        }
+        self.address = value
+    }
+    
+    func getAddress() -> String {
+        return self.address
+    }
+}
+
+enum errorAddress: Error {
+    case empty
+}
+
+struct Name {
+    private let name: String
+    init(value: String) throws {
+        guard !value.isEmpty else {
+            throw errorName.empty
+        }
+        self.name = value
+    }
+    
+    func getName() -> String {
+        return self.name
+    }
+}
+
+enum errorName: Error {
+    case empty
+}
+
+struct Mobile {
+    private let mobile: String
+    
+    init(value: String) throws {
+        try Self.validate(value: value)
+        self.mobile = value
+    }
+    
+    func getMobile() -> String {
+        return self.mobile
+    }
+    
+    private static func validate(value: String) throws {
+        guard !value.isEmpty else {
+            throw errorMobile.empty
+        }
+        guard value.count < 13 else {
+            throw errorMobile.rangeOver
+        }
+    }
+}
+
+enum errorMobile: Error {
+    case empty
+    case rangeOver
+}
+
+struct Person {
+    private let address: Address
+    private var name: Name
+    private var mobile: Mobile
+    
+    init(address: Address, name: Name, mobile: Mobile) {
+        self.address = address
+        self.name = name
+        self.mobile = mobile
+    }
+    
+    func getAddress() -> String {
+        return self.address.getAddress()
+    }
+    func getName() -> String {
+        return self.name.getName()
+    }
+    func getMobile() -> String {
+        return self.mobile.getMobile()
+    }
+}
